+++ reader.h
@@ -12,7 +12,7 @@ public:
 	bool Initialize(std::string);//gives the reader a file and it checks the validity of said file. Will output false if it can't be opened
 	std::vector<Molecule> GetMoleculesFromDump();
 	Atom ParseAtomLine(std::string line);
-	float ScientificNotationToFloat(std::string numberAsString);
-	float GetCubeSizeFromLine(std::string line);
+	double ScientificNotationToFloat(std::string numberAsString);
+	double GetCubeSizeFromLine(std::string line);
 	Molecule GetMoleculeFromOutputFile();
 };
